requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
contourpy==1.3.1; python_version >= '3.10'
cycler==0.12.1; python_version >= '3.8'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==4.1.13; python_version >= '3.8'
django-cors-headers; python_version >= '3.9'
djangorestframework; python_version >= '3.8'
djongo==1.3.6; python_version >= '3.6'
dnspython==2.7.0; python_version >= '3.9'
filelock==3.17.0; python_version >= '3.9'
filetype==1.2.0
fonttools==4.56.0; python_version >= '3.8'
fsspec==2025.2.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
jinja2==3.1.5; python_version >= '3.7'
joblib==1.4.2; python_version >= '3.8'
kiwisolver==1.4.8; python_version >= '3.10'
markupsafe==3.0.2; python_version >= '3.9'
matplotlib==3.10.0; python_version >= '3.10'
mpmath==1.3.0
networkx==3.4.2; python_version >= '3.10'
numpy; python_version >= '3.10'
opencv-python==4.11.0.86; python_version >= '3.6'
opencv-python-headless==4.10.0.84; python_version >= '3.6'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pillow==11.1.0; python_version >= '3.9'
psutil==7.0.0; python_version >= '3.6'
py-cpuinfo==9.0.0
pymongo==3.12.3
pyparsing==3.2.1; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2025.1
pyyaml==6.0.2; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
roboflow==1.1.53; python_version >= '3.8'
scikit-learn==1.6.1; python_version >= '3.9'
scipy==1.15.2; python_version >= '3.10'
seaborn==0.13.2; python_version >= '3.8'
setuptools==75.8.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.2.4
supervision==0.25.1; python_version >= '3.8' and python_version < '4.0'
sympy==1.13.1; python_version >= '3.8'
threadpoolctl==3.5.0; python_version >= '3.8'
torch==2.6.0; python_full_version >= '3.9.0'
torchvision==0.21.0; python_version >= '3.9'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2025.1; python_version >= '2'
ultralytics==8.3.40; python_version >= '3.8'
ultralytics-thop==2.0.14; python_version >= '3.8'
urllib3==2.3.0; python_version >= '3.9'
langchain==0.3.20; python_version >= '3.9'
langchain-groq==0.2.5; python_version >= '3.9'
langchain-openai==0.3.8; python_version >= '3.9'
langchain-anthropic==0.3.9; python_version >= '3.9'
langchain-google-vertexai==2.0.15
langchain-mongodb==0.5.0